Move compression functions into helper module
